+++ src/hooks/useFetchCategoryPoses.tsx
@@ -2,7 +2,8 @@ import { useQuery } from '@tanstack/react-query';
 import CATEGORY from '../types/CategoryType';
 import { PoseInfo } from '../types/PoseType';
 import { ALL_PEOPLE_TAG } from '../constant/tagId';
-import { poseService } from '../api/poseService';
+import { apiService } from '../service/ApiService';
+import { useFetchAllPose } from '../queries/poses';
 
 export default function useFetchCategoryPoses(
   category: CATEGORY,
@@ -10,15 +11,27 @@ export default function useFetchCategoryPoses(
 )
   :PoseInfo[] {
   if (category === CATEGORY.PEOPLE && selectedTagId === ALL_PEOPLE_TAG) {
-    const { data } = useQuery({ queryKey: ['allPoses'], queryFn: () => poseService.fetchAllPoses() });
+    const { data } = useFetchAllPose();
     return data ?? [];
   }
   if (category === CATEGORY.PEOPLE || category === CATEGORY.THEME) {
-    const { data } = useQuery({ queryKey: ['pose', selectedTagId], queryFn: () => poseService.fetchPoses([selectedTagId]) });
+    const { data } = useQuery({
+      queryKey: ['pose', selectedTagId],
+      queryFn: async () => {
+        const { data: poses } = await apiService.fetchPose([selectedTagId]);
+        return poses;
+      },
+    });
     return data ?? [];
   }
   if (category === CATEGORY.POPULAR) {
-    const { data } = useQuery({ queryKey: ['popularPoses'], queryFn: () => poseService.fetchPopularPoses() });
+    const { data } = useQuery({
+      queryKey: ['popularPoses'],
+      queryFn: async () => {
+        const { data: poses } = await apiService.fetchPopularPose();
+        return poses;
+      },
+    });
     return data ?? [];
   }
   return [];
