+++ src/hooks/useFetchMyPoses.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { authPoseService } from '../api/authPoseService';
+import { authPoseService } from '../service/AuthPoseService';
 
 export default function useFetchMyPoses() {
   return useQuery({
