+++ src/service/ApiService.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { ResponseFetchPoses, PoseType } from '../types/PoseType';
+import { ResponseFetchPoses } from '../types/PoseType';
 
 const API_BASE_URL = `${process.env.REACT_APP_API_BASE_URL}/api/v1` || 'https://server.poseplz.com/api/v1';
