+++ src/types/PoseType.ts
@@ -9,15 +9,6 @@ export interface File {
   width: number;
   height: number;
 }
-export interface PoseType {
-  poseId: string;
-  thumbnailImageUrl: string;
-  imageUrl: string;
-  file: File;
-  tags: Tag[];
-  peopleCount: number;
-  archived: boolean;
-}
 
 export interface PoseInfo {
   poseId: string;
@@ -35,10 +26,12 @@ export interface Pagination {
   page: number;
   size: number;
 }
+
 export interface ResponseFetchPoses {
   data: PoseInfo[];
   pagination: Pagination;
 }
+
 export interface MyPoseType {
   poseId: string;
   thumbnailImageUrl: string;
